DB의 저장된 파일 및 usb 대출가능 여부 및 프로그램 목록 업로드

+++ src/main/java/com/tak/USB/Manager/repository/ProgramRepository.java
@@ -3,5 +3,5 @@ package com.tak.USB.Manager.repository;
 import com.tak.USB.Manager.entity.ProgramEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProgramRepository extends JpaRepository<ProgramEntity, Integer> {
+public interface ProgramRepository extends JpaRepository<ProgramEntity, Long> {
 }
